+++ src/others/ft_isalnum.c
@@ -12,7 +12,7 @@
 
 #include "../include/libft.h"
 
-int	ft_isalnum(int c)
+int8_t	ft_isalnum(int c)
 {
 	return (ft_isalpha(c) || ft_isdigit(c));
 }
